+++ AirsoftBmsApp/Model/Dto/Kills/PostKillDto.cs
@@ -4,7 +4,7 @@ public class PostKillDto
 {
     public double Longitude { get; set; }
     public double Latitude { get; set; }
-    public double Accuracy { get; set; }
-    public double Bearing { get; set; }
+    public short Accuracy { get; set; }
+    public short Bearing { get; set; }
     public DateTimeOffset Time { get; set; }
 }
